add overlay shortcut for window
Adds a Alt+Shift+N shortcut that toggles the visibility of the app, and makes the app visible on all workspaces. This gives the user access to write notes with a single keystroke, and then hide the notes and get back to work by focusing the previous window.

There are notes in this commit for future work to be done, by adding a true "overlay" mode that removes the app from the dock and alt/cmd+tab menu and instead adds menubar access to a taskbar icon.